pyproject.toml
[tool.poetry]
name = "dashcv"
version = "0.1.0"
description = ""
authors = ["yannick.schwarz <yannick.schwarz@kvg.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.15.0"
dash-bootstrap-components = "^1.5.0"
pandas = "^2.2.0"
pip = "^24.0"
py-plot-ge = {git = "https://github.com/yaennu/py_plot_ge"}
nbformat = "^5.9.2"
python-dotenv = "^1.0.1"
pymssql = "^2.2.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
